Show kingdom as a summary on main page
This changes Kingdom into an actual class where we can get useful data.
This will become more and more complex over time, I hope. For now I just
wanted to show a compressed view of the kingdom on the summary page with
a link to a more detailed view to make for easier browsing