Add working rating widget
This includes the ability to rate a kingdom by clicking on those cool
stars that change as you hover over them. Yay.

I changed a bunch about the javascript and made dominus.js load at the
end of the page so I could start adding jQuery style setup where I take
some HTML and make it more awesome. This went along well with the star
rater I got from rog.ie:

http://rog.ie/blog/css-star-rater

I now have a dependency on less I'm not crazy about, but I'll live with
it until it drives me crazy enough to check in the CSS it generates or I
convert it to something like SASS that I'm more okay with